VRC: More ints are unsigned and need updated format strings.

+++ frmts/vrc/VRCthirtysix.cpp
@@ -79,7 +79,9 @@ void VRCRasterBand::read_VRC_Tile_36(VSILFILE *fp, int block_xx, int block_yy,
     CPLDebug("Viewranger", "read_VRC_Tile_36(%p, %d, %d, %p)",
              static_cast<void *>(fp), block_xx, block_yy, pImage);
 
-    const int tilenum = poGDS->tileXcount * block_yy + block_xx;
+    const unsigned int tilenum =
+        poGDS->tileXcount * static_cast<unsigned int>(block_yy) +
+        static_cast<unsigned int>(block_xx);
     // VRC36_PIXEL_IS_PIXEL
     // this will be the default
     CPLDebug("Viewranger", "vrc36_pixel_is_pixel only partially implemented");
